backend/requirements.txt
# Core Framework
fastapi[standard]==0.128.8
uvicorn[standard]==0.40.0
python-multipart==0.0.20

# Data Validation
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.7.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# Environment
python-dotenv==1.0.0

# Database & ORM
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
faker==33.1.0